% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.glm.R
\name{backward.glm}
\alias{backward.glm}
\title{Backward variable selection using given full model}
\usage{
backward.glm(model, sls = 0.05, printing = TRUE)
}
\arguments{
\item{model}{A full 'glm' model preliminarily built with formula/data/family parameters only}

\item{sls}{A significance level for stay (sls) applied for each backward elimination step}

\item{printing}{A logical parameter indicating report information to console output}
}
\value{
The shrunk 'glm' model where all input variables satisfy significance level for stay
}
\description{
Takes in any full 'glm' model and significance level for stay (sls) then conducts backward variable selection and returns shrunk model.
}
\examples{
# Example 1:
library(mlbench)
data(PimaIndiansDiabetes2)
df <- na.omit(PimaIndiansDiabetes2) # removing rows with any NA values
df$diabetes <- ifelse(df$diabetes == "pos", 1, 0)
full_model <- glm(diabetes ~ ., data = df, family = binomial())
print(formula(full_model))
library(StepwiseBinaryLogisticRegression)
model <- backward.glm(full_model)
print(formula(model))
print(summary(model))

# Example 2:
library(StepwiseBinaryLogisticRegression)
data(Remission)
full_model2 <- glm(remiss ~ ., data = Remission, family = binomial())
print(formula(full_model2))
model2 <- backward.glm(full_model2, sls = 0.35)
print(formula(model2))
print(summary(model2))
}
