% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entry_frame.R
\name{get_entry_frame}
\alias{get_entry_frame}
\title{Estimation of all significance levels for entry to binary logistic regression model using current model and external input variables}
\usage{
get_entry_frame(complete_data_frame, current_model, entry_name_vector)
}
\arguments{
\item{complete_data_frame}{A data frame containing input variables from the current model and all to be tested external numeric input variables (without NA values) which exist in entry name vector}

\item{current_model}{A current binary logistic regression model}

\item{entry_name_vector}{Entry name vector containing names of external numeric input variables (without NA values) to be tested}
}
\value{
The data frame containing test results for entry of external input variables which includes external names, chi-squared values and p-values for entry
}
\description{
Takes in data frame containing both currently used input variables and external input variables also taking in current binary logistic regression model and names of external input variables; returns result data frame with external names, chi-squared values and corresponding p-values for entry.
}
\examples{
# Example 1:
library(mlbench)
data(PimaIndiansDiabetes2)
df <- na.omit(PimaIndiansDiabetes2)
df$diabetes <- ifelse(df$diabetes == "pos", 1, 0)
m <- glm(diabetes ~ glucose + age + mass + pedigree, df, family = binomial())
library(StepwiseBinaryLogisticRegression)
print(get_entry_frame(df, m, c("pregnant", "pressure", "triceps", "insulin")))

# Example 2:
library(StepwiseBinaryLogisticRegression)
data(Remission)
m2 <- glm(remiss ~ li + temp + cell, Remission, family = binomial())
print(get_entry_frame(Remission, m2, c("smear", "infil", "blast")))
}
