% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Remission}
\alias{Remission}
\title{Cancer remission data}
\format{
An object of class data.frame with 27 rows and 7 columns:
\describe{
\item{cell}{The 1st explanatory variable "x1" (x1 = CELL/100)}
\item{smear}{The 2nd explanatory variable "x2" (x2 = SMEAR/100)}
\item{infil}{The 3rd explanatory variable "x3" (x3 = ABS.INFIL/100)}
\item{li}{The 4th explanatory variable "x4" (x4 = LI/20)}
\item{blast}{The 5th explanatory variable "x5" (x5 = log(ABS.BLAST+1))}
\item{temp}{The 6th explanatory variable "x6" (x6 = Temperature/100)}
\item{remiss}{The binary response variable where 1 implies that the patient’s remission is occured and 0 otherwise}
}
}
\usage{
Remission
}
\description{
Cancer remission data
}
\references{
Tan, M. T., Tian, G.-L., & Ng, K. W. (2010). \emph{Bayesian Missing Data Problems: EM, Data Augmentation and Noniterative Computation.} Chapman & Hall/CRC.

Lee, E. T. (1974). “A Computer Program for Linear Logistic Regression Analysis.” Computer Programs in Biomedicine 4:80–92.
}
\keyword{datasets}
