% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepwiseBinaryLogisticRegression.R
\name{StepwiseBinaryLogisticRegression}
\alias{StepwiseBinaryLogisticRegression}
\title{Creation of binary logistic regression model using stepwise variable selection with significance levels for entry (sle) and for stay (sls)}
\usage{
StepwiseBinaryLogisticRegression(
  formula,
  data,
  sle = 0.05,
  sls = 0.05,
  printing = TRUE
)
}
\arguments{
\item{formula}{A formula which contains both binary response variable and not empty numeric variables (without NA values) separated with '+' sign or just '.' sign to try all not empty numeric inputs (without NA values). An example: \code{y ~ .} or \code{y ~ x1 + x2 + x3}}

\item{data}{A data frame containing both binary response variable and not empty (without NA values) numeric input variables (which are mentioned in formula)}

\item{sle}{A significance level for entry (for each forward entry step)}

\item{sls}{A significance level for stay (for each backward elimination step)}

\item{printing}{A logical parameter indicating report information to console output}
}
\value{
The binomial/binary 'glm' model with logit link function containing only the input variables included with stepwise variable selection
}
\description{
Takes in formula with both binary response variable and not empty numeric variables separated with '+' sign or just '.' sign to try all numeric not empty inputs also taking in data frame and significance levels for entry and for stay.
}
\examples{
# Example 1:
library(mlbench)
data(PimaIndiansDiabetes2)
df <- na.omit(PimaIndiansDiabetes2) # removing rows with any NA values
df$diabetes <- ifelse(df$diabetes == "pos", 1, 0)
library(StepwiseBinaryLogisticRegression)
m <- StepwiseBinaryLogisticRegression(diabetes ~ pregnant + glucose + pressure +
triceps + insulin + mass + pedigree + age, df)
print(formula(m))
print(summary(m))

# Example 2 (actually the same one in this situation):
library(mlbench)
data(PimaIndiansDiabetes2)
df <- na.omit(PimaIndiansDiabetes2) # removing rows with any NA values
df$diabetes <- ifelse(df$diabetes == "pos", 1, 0)
library(StepwiseBinaryLogisticRegression)
m2 <- StepwiseBinaryLogisticRegression(diabetes ~ ., df)
print(formula(m2))
print(summary(m2))

# Example 3:
library(StepwiseBinaryLogisticRegression)
data(Remission)
# here sle <= sls just in case to prevent any possible looping:
m3 <- StepwiseBinaryLogisticRegression(remiss ~ ., Remission, sle = 0.30, sls = 0.35)
print(formula(m3))
print(summary(m3))
}
